requirements.txt
Flask
Pillow
google-generativeai
beautifulsoup4
requests
python-dotenv